scripts/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
openai>=1.0.0
plotly>=5.14.0
scipy>=1.10.0
scikit-learn>=1.2.0
streamlit>=1.25.0
pyyaml>=6.0.0
tqdm>=4.65.0